+++ templates/viewcontent.php
@@ -9,43 +9,76 @@ echo '<p id="gpxmotion_version" style="display:none">';
 p($_['gpxmotion_version']);
 echo '</p>'."\n";
 
-echo '<ul id="overlayserverlist" style="display:none">';
-foreach($_['overlayservers'] as $name=>$url){
-    echo '<li name="';
-    p($name);
-    echo '" title="';
-    p($url);
-    echo '">';
-    echo '</li>';
-}
+echo '<ul id="tileserverlist" style="display:none">';
+    foreach($_['usertileservers'] as $ts){
+        echo '<li title="'.$ts['url'].'"';
+        foreach (Array('servername', 'type', 'url', 'layers', 'version', 'format', 'opacity', 'transparent', 'minzoom', 'maxzoom', 'attribution') as $field) {
+            if (array_key_exists($field, $ts)) {
+                echo ' '.$field.'="';
+                p($ts[$field]);
+                echo '"';
+            }
+        }
+        echo '>';
+        echo '</li>';
+    }
 echo '</ul>'."\n";
 
-echo '<ul id="tileserverlist" style="display:none">';
-foreach($_['tileservers'] as $name=>$url){
-    echo '<li name="';
-    p($name);
-    echo '" title="';
-    p($url);
-    echo '">';
-    echo '</li>';
-}
+echo '<ul id="overlayserverlist" style="display:none">';
+    foreach($_['useroverlayservers'] as $ts){
+        echo '<li title="'.$ts['url'].'"';
+        foreach (Array('servername', 'type', 'url', 'layers', 'version', 'format', 'opacity', 'transparent', 'minzoom', 'maxzoom', 'attribution') as $field) {
+            if (array_key_exists($field, $ts)) {
+                echo ' '.$field.'="';
+                p($ts[$field]);
+                echo '"';
+            }
+        }
+        echo '>';
+        echo '</li>';
+    }
+echo '</ul>'."\n";
+
+echo '<ul id="overlaywmsserverlist" style="display:none">';
+    foreach($_['useroverlayserverswms'] as $ts){
+        echo '<li title="'.$ts['url'].'"';
+        foreach (Array('servername', 'type', 'url', 'layers', 'version', 'format', 'opacity', 'transparent', 'minzoom', 'maxzoom', 'attribution') as $field) {
+            if (array_key_exists($field, $ts)) {
+                echo ' '.$field.'="';
+                p($ts[$field]);
+                echo '"';
+            }
+        }
+        echo '>';
+        echo '</li>';
+    }
+echo '</ul>'."\n";
+echo '<ul id="tilewmsserverlist" style="display:none">';
+    foreach($_['usertileserverswms'] as $ts){
+        echo '<li title="'.$ts['url'].'"';
+        foreach (Array('servername', 'type', 'url', 'layers', 'version', 'format', 'opacity', 'transparent', 'minzoom', 'maxzoom', 'attribution') as $field) {
+            if (array_key_exists($field, $ts)) {
+                echo ' '.$field.'="';
+                p($ts[$field]);
+                echo '"';
+            }
+        }
+        echo '>';
+        echo '</li>';
+    }
 echo '</ul>'."\n";
 
 echo '<ul id="basetileservers" style="display:none">';
 foreach($_['basetileservers'] as $ts){
-    echo '<li name="';
-    p($ts['name']);
-    echo '" type="';
-    p($ts['type']);
-    echo '" url="';
-    p($ts['url']);
-    echo '" minzoom="';
-    p($ts['minzoom']);
-    echo '" maxzoom="';
-    p($ts['maxzoom']);
-    echo '" attribution="';
-    p($ts['attribution']);
-    echo '"></li>';
+    echo '<li';
+    foreach (Array('name', 'type', 'url', 'layers', 'version', 'format', 'opacity', 'transparent', 'minzoom', 'maxzoom', 'attribution') as $field) {
+        if (array_key_exists($field, $ts)) {
+            echo ' '.$field.'="';
+            p($ts[$field]);
+            echo '"';
+        }
+    }
+    echo '></li>';
 }
 echo '</ul>'."\n";
 ?>